requirements.txt
openai
line-bot-sdk
boto3